% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{assist_transform_addin}
\alias{assist_transform_addin}
\title{Low Code Assistant - Transformations}
\usage{
assist_transform_addin()
}
\description{
Run the data transformation LCA wizard. If a variable holding a data frame is selected
when invoking the addin in RStudio, that data frame will be used.
}
