% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{assist_viz_addin}
\alias{assist_viz_addin}
\title{Low Code Assistant - Visualizations}
\usage{
assist_viz_addin()
}
\description{
Run the data visualizations LCA wizard. If a variable holding a data frame is selected
when invoking the addin in RStudio, that data frame will be used.
}
