% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-chunk.R
\name{code_chunk_ui}
\alias{code_chunk_ui}
\title{Code chunk module}
\usage{
code_chunk_ui(id)
}
\description{
Display a series of code chunks with syntax highlighting. Each chunk can optionally
have Insert/Modify/Delete buttons. One chunk can be shown as an error/bug.
}
