% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-xforms-table.R
\name{xforms_table}
\alias{xforms_table}
\alias{xforms_table_ui}
\alias{xforms_table_server}
\title{Transformations table}
\usage{
xforms_table_ui(id)

xforms_table_server(id, data)
}
\arguments{
\item{id}{Unique ID for the module}

\item{data}{(reactive) Data to show in the table}
}
\value{
List:
\itemize{
\item drop: (reactive) A drop transformation that the user selected
\item missing: (reactive) A missing values transformation that the user selected
\item filter: (reactive) A filter transformation that the user selected (the 'type' isn't set correctly)
}
}
\description{
Transformations table
}
