% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{assist_transform}
\alias{assist_transform}
\title{Low Code Assistant - Transformations}
\usage{
assist_transform(data_name = NULL)
}
\arguments{
\item{data_name}{The name of a data frame in the global environment. Can be reactive or static.}
}
\description{
Run the data transformation LCA wizard.
}
